Reorganized frontend with Storybook 10 + module mocking so Windows and Apps render without the backend
