+++ frontend/package.json
@@ -2,6 +2,28 @@
     "name": "proompting-party",
     "private": true,
     "type": "module",
+    "imports": {
+        "#api/party-zone": {
+            "storybook": "./src/api/party-zone.mock.ts",
+            "default": "./src/api/party-zone.ts"
+        },
+        "#lib/realtime-store": {
+            "storybook": "./src/lib/realtime-store.mock.ts",
+            "default": "./src/lib/realtime-store.ts"
+        },
+        "#lib/desktop-context": {
+            "storybook": "./src/lib/desktop-context.mock.tsx",
+            "default": "./src/lib/desktop-context.tsx"
+        },
+        "#lib/nav-history": {
+            "storybook": "./src/lib/nav-history.mock.ts",
+            "default": "./src/lib/nav-history.ts"
+        },
+        "#lib/persona-ws-stream": {
+            "storybook": "./src/lib/persona-ws-stream.mock.ts",
+            "default": "./src/lib/persona-ws-stream.ts"
+        }
+    },
     "scripts": {
         "dev": "vite dev",
         "build": "vite build",
@@ -10,7 +32,9 @@
         "format": "biome format --write",
         "lint": "biome lint --write",
         "check": "biome check",
-        "api-generate": "orval"
+        "api-generate": "orval",
+        "storybook": "storybook dev -p 6006",
+        "build-storybook": "storybook build"
     },
     "dependencies": {
         "@tailwindcss/vite": "^4.2.2",
@@ -34,12 +58,12 @@
         "tailwindcss": "^4.1.18",
         "valibot": "^1.2.0",
         "vite": "^8.0.7",
-        "vite-tsconfig-paths": "^5.1.4",
         "xp.css": "^0.2.6",
         "zustand": "^5.0.11"
     },
     "devDependencies": {
         "@biomejs/biome": "2.2.4",
+        "@storybook/react-vite": "^10.3.5",
         "@tanstack/devtools-vite": "0.6.0",
         "@tanstack/react-router-ssr-query": "^1.166.10",
         "@types/node": "^25.6.0",
@@ -48,6 +72,7 @@
         "@types/react-grid-layout": "^1.3.5",
         "@vitejs/plugin-react": "^6.0.1",
         "orval": "^8.5.3",
+        "storybook": "^10.3.5",
         "typescript": "^5.7.2"
     }
 }
