+++ frontend/src/lib/nav-history.mock.ts
@@ -1,0 +1,1 @@
+export * from './nav-history';
