+++ frontend/src/lib/realtime-store.ts
@@ -493,8 +493,7 @@ export const useRealtimeStore = create<RealtimeStoreState>((set, get) => {
                         /* ignore */
                     }
                     const declined: DeclinedDecision = {
-                        personaId:
-                            declinedPersonaId || phantom?.senderId || '',
+                        personaId: declinedPersonaId || phantom?.senderId || '',
                         reason: declinedReason,
                         decisionText: (phantom?.appraisal ?? '').trim(),
                         timestamp: Date.now(),
