+++ frontend/src/router.tsx
@@ -1,7 +1,7 @@
+import { QueryClient } from '@tanstack/react-query';
 import { createRouter, ErrorComponent } from '@tanstack/react-router';
 import { setupRouterSsrQueryIntegration } from '@tanstack/react-router-ssr-query';
 import { routeTree } from './routeTree.gen';
-import { QueryClient } from '@tanstack/react-query';
 
 export function getRouter() {
     const queryClient = new QueryClient();
