+++ frontend/src/server.ts
@@ -1,5 +1,6 @@
 import handler, { createServerEntry } from '@tanstack/react-start/server-entry';
 import { FastResponse } from 'srvx';
+
 globalThis.Response = FastResponse;
 
 const serverEntry = createServerEntry({
