+++ frontend/src/views/DesktopExperience.tsx
@@ -1,4 +1,4 @@
-import { useMemo, createElement } from 'react';
+import { createElement, useMemo } from 'react';
 import GridLayout from 'react-grid-layout';
 import 'react-grid-layout/css/styles.css';
 import DesktopIcon from '../components/desktop/DesktopIcon';
